requirements.txt
django >=1.9, <2
viberbot==1.0.11
django-import-export==1.0.0